requirements.txt
streamlit==1.28.1
openai>=1.6.1,<2.0.0
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5,<0.1.0
PyMuPDF>=1.23.8
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0 